requirements.txt
streamlit>=1.28.0
pathlib2>=2.3.7
plotly>=5.15.0
websockets>=11.0
pydantic>=2.0.0
python-dotenv==1.1.0
langgraph==0.3.1
